port(Lifestream): TC API12 source (based on TC_forward @ 24d02186)

+++ Lifestream/IPC/Wotsit/WotsitManager.cs
@@ -113,7 +113,7 @@ public class WotsitManager : IDisposable
         MaybeTryInit(false);
     }
 
-    private void TerritoryChanged(uint territory)
+    private void TerritoryChanged(ushort territory)
     {
         PluginLog.Debug($"WotsitManager: Territory changed to {territory}, attempting re-registration");
         MaybeTryInit(false);
